requirements.txt
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
asgiref==3.4.1
astroid==2.9.3
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
chardet==4.0.0
charset-normalizer==2.0.9
click==8.1.3
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
daphne==3.0.2
decorator==5.1.1
defusedxml==0.7.1
dicttoxml==1.7.4
Django==3.2
django-allauth==0.47.0
django-elasticsearch-dsl==7.2.0
django-elasticsearch-dsl-drf==0.22.4
django-email-users==1.1.0
django-ipware==4.0.2
django-nine==0.2.5
django-pwa==1.0.10
djangorestframework==3.13.1
docType==0.23
docutils==0.18.1
elasticsearch==7.13.4
elasticsearch-dsl==7.4.0
frozenlist==1.3.0
future==0.18.2
geocoder==1.38.1
geographiclib==1.52
geoip2==4.5.0
geopy==2.2.0
h11==0.13.0
h3==3.7.3
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.10.1
incremental==21.3.0
install==1.3.5
ip2geotools==0.1.6
IP2Location==8.6.4
ipaddress==1.0.23
isort==5.10.1
joblib==1.1.0
keyring==23.5.0
lazy-object-proxy==1.7.1
lxml==4.7.1
maxminddb==2.2.0
mccabe==0.6.1
msgpack==1.0.3
multidict==5.2.0
nltk==3.7
numpy==1.22.4
oauthlib==3.1.1
outcome==1.1.0
packaging==21.3
pandas==1.4.1
pgeocode==0.3.0
Pillow==9.0.1
pip-review==1.1.1
pkginfo==1.8.2
platformdirs==2.4.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pylint==2.12.2
pyOpenSSL==21.0.0
pyparsing==3.0.7
pyquery==1.4.3
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
ratelim==0.1.6
readme-renderer==32.0
regex==2022.6.2
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==2.0.0
s2sphere==0.2.5
scikit-learn==1.1.1
scipy==1.8.1
selenium==4.1.0
service-identity==21.1.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.64.0
trio==0.19.0
trio-websocket==0.9.2
twine==3.7.1
Twisted==22.1.0
txaio==21.2.1
typed-ast==1.5.1
typing==3.7.4.3
typing_extensions==4.0.1
urllib3==1.26.7
webencodings==0.5.1
wrapt==1.13.3
wsproto==1.0.0
yarl==1.7.2
zipp==3.7.0
zope.interface==5.4.0
